requirements.txt
requests
openai[azure]
python-dotenv
gitpython
pydantic
jira 